reviewboard: disallow pushing and pulling discovery repos (bug 1189575); r=smacleod
Attempting to push to a discovery repository without the reviewboard
client extension installed results in a really bad experience. The
Mercurial client will see the discovery repo as if its a normal
Mercurial repository (no problem with that). However, it will then see
it is empty and attempt to upload all local changesets to it. This
wouldn't be bad except Mercurial first bundles all changesets to a
temporary local file then uploads the bundle in full before the server
applies it. You can imagine what this is like on mozilla-central :/

This patch short circuits the "heads" wire protocol command on discovery
repos. This command is called as part of changeset discovery so clients
know what to send and receive from the server. By making it error, we
quickly fail all push and pull operations. We print an actionable error
message so people can rectify the problem by installing the appropriate
Mercurial extension.