mozreview: only mirror full commit description to bugzilla when first publishing (bug 1216078) r=mcote
At the moment we create a comment containing the full commit description each
time a commit is updated. If people write detailed commit messages, this ends
up taking a large amount of space when looking at the bug.

With these changes, we only create a comment with the full commit description
on the first change. The comment for each subsequent update to a commit only
contains a link to the interdiff.