push-to-try: always release transaction object (bug 1216436); r=chmanchester
The proper pattern for transactions is:

tr = repo.transaction(...)
try:
   ...
   tr.close() # or tr.abort()
finally:
   tr.release()

We were forgetting the tr.release() in the finally block, meaning that
the transaction could have not been rolled back.

Surprisingly, I was unable to reproduce the reported bug. But I've added
an explicit test just so we have test coverage.