bundleclone: make tests pass on old Mercurial versions; r=smacleod
The tests for bundle manifest copying were failing on Mercurial releases
older than 3.0. We don't care about this functionality on these old
versions. So just don't run the tests there.

Since we added the first consumer of hghave, we needed to install a stub
file in the test directory to make the test harness happy.