bundleclone: disable bundleclone when clone bundles feature detected (bug 1216216); r=smacleod
Mercurial 3.6 has official support for the bundle clone feature set. The
bundleclone extension is thus redundant with that functionality (or at
least it should if the server operator is competent).

This commit adds detection of the clone bundles feature to the client
and will short-circuit the bundleclone functionality when the built-in
support for cloning from pre-generated bundles is sufficient.