bundleclone: support producing clone bundle compatible stream bundles (bug 1216216); r=smacleod
When bundleclone was initially implemented, the concept of a "stream
bundle" was invented. It was essentially the wire protocol
representation of the existing streaming clone data saved to a file.

Mercurial 3.6 formalized the "stream bundle" format. It is slightly
different from and incompatible with the wire protocol format.

This patch teaches our `hg streambundle` command to produce the new,
formal format. We also add tests verifying a client requests the
appropriate file depending on whether clone bundles or bundleclone is
used.