scripts/generate-hg-s3-bundles: create clonebundles.manifest (bug 1216216); r=smacleod
The built-in Mercurial feature for serving pre-built bundles is slightly
different from bundleclone. It uses a separate manifest that has a
slightly different format (mostly cosmetic changes).

With this commit, we start writing this new manifest file when
generating bundles. As part of this, we generate the modern stream
bundle file format.