pushlog: don't use weakref to repo (bug 1217569); r=smacleod
The old behavior won't work in Mercurial 3.6 because of
https://bz.mercurial-scm.org/show_bug.cgi?id=4915.

This is arguably a bug in Mercurial. However, looking at the Mercurial
code, relying on the same exact object for the repo is kinda difficult,
as filters are involved and these create new repo instances.

This commit removes the weakref. It also stops caching the pushlog
instance on the repo. There is no state on the instance and recreating
it is cheap. We arguably should have been doing this all along.

Since this was the last piece of pushlog that didn't work with 3.6, I
bumped the testwith property to indicate 3.6 compatibility.