mozreview: convert nl to br in try error message (bug 1175759); r=dminor
The changedescription object corresponding to an autoland to try failure
will render \n characters as <br /> to maintain the original formatting.
In case of a very long error message this shouldn't be a problem as the
changedescription panels can be collapsed.
The autoland panel at the bottom of the commit table cannot be collapsed,
so I decided to only show the first line of content and place the full
error message in the title of the span containing it..