ansible/hgmo: don't advertise clone bundles feature (bug 1221268); r=fubar
clonebundles has a pretty severe bug in 3.6 that causes it to attempt to
apply a clone bundle during `hg pull`. Count this as fallout from me
being on PTO and not testing this feature before 3.6 landed.
https://bz.mercurial-scm.org/show_bug.cgi?id=4932

The upstream fix has already been accepted and will be in 3.6.1, which
will be released December 1 unless there is a regression that prompts a
release sooner.