hgmo: replace bookmarks when replicating (bug 1139056); r=dminor
Currently, the mirroring logic performs a `hg pull`. The default
behavior of `hg pull` w.r.t. bookmarks is somewhat complicated. If a
bookmark was not fast-forwarded, it is diverged. There is currently no
knob in Mercurial to change this behavior (although remotenames will add
it eventually). This results in bookmarks like "@default" appearing and
the bookmark state on hgweb machines getting out of sync with the
master.

This commit introduces a wrapping of the bookmarks function in the hgmo
extension. When the "hgmo.replacebookmarks" config flag is set, we
replace the built-in updating logic with a simple "replace from remote"
scheme.

We modify our replication script on the hgweb machines to define this
config flag when performing `hg pull`. As a result, we no longer have
bookmark divergence on hgweb machines and bookmarks work as they should.

A test demonstrating old, incorrect behavior has been modified to
reflect the new, proper behavior. New tests verifying that bookmark
deletion work have been added.