hghooks: advertise mercurial upgrade to non-bundle2 clients (bug 1221827); r=dminor
Mercurial 3.5 and newer support a new wire protocol exchange format
called "bundle2." Not using bundle2 almost certainly means the client is
on an old version of Mercurial.

In an attempt to nudge people into using a modern version of Mercurial,
this commit adds a hook that will print a message to non-bundle2 users
during push encouraging them to upgrade. The message is purely advisory
and we have no plans to lock out legacy clients at this time.