hghooks: add hook to validate author field well-formedness (bug 787620); r=glandium
Mercurial isn't strict about what goes in the author/user field. It is
supposed to be an RFC 822 "NAME <EMAIL>" value. However, Mercurial by
itself will allow pretty much everything. Historically, this has caused
problems converting Mercurial repos to Git repos, as Git is more strict
about what goes in this field. The tools do normalize values to
something Git will accept. But it would be nice to avoid the problem in
the first place so there is parity between Mercurial and Git repos.

This commit introduces a new Mercurial hook that validates the
well-formedness of the author/user field on changesets. The hook has
been globally enabled on hg.mozilla.org.