readonly: add an extension that marks repositories as read-only (bug 1183282); r=dminor
Lots of repositories on hg.mozilla.org are ancient and should no longer
be updated. In addition, it is desirable during some server
maintenance operations to prevent all pushes.

This commit introduces a new "readonly" extension that can be used to
mark repositories as read only. You can put a file in the .hg/ of an
individual repository to mark it read only. Or you can configure a
global read-only file to be used to mark all repositories as read only.
This satisfies both our major use cases.

It is unfortunate we can't hook into pretxnopen. However, the Mercurial
bug not showing output from this hook is a deal breaker :/