ansible: run a Kafka cluster on Mercurial servers
This commit add a kafka-broker Ansible role that is used to configure a
ZooKeeper and Kafka server on a machine. We add the role to the hg-web
and hg-ssh roles so ZooKeeper and Kafka run from our Mercurial servers.

We update our hg.mozilla.org test cluster to support starting ZooKeeper
and Kafka clusters that talk to each Docker container.

The configuration for hg.mozilla.org has been updated to start a
ZooKeeper and Kafka cluster for all servers.