vcsreplicator: ability to replicate hgrc files
We now have the ability to record and apply the contents of hgrc
files in the replication log. hgrc change events are written through
the `hg replicatehgrc` command.

Mercurial does provide a `hg config -e` mode that writes configs. We
could potentially hook into that. For now, the command is sufficient.