vcsreplicator: record pushkey changes
Mercurial tracks data like bookmarks, phases, and obsolescence via the
"pushkey" protocol. It is essentially a namespace of key-value pairs.

We intercept the hook that fires when pushkey entries are changed and
produce a message containing the changes.

When we receive these messages, we call into `hg debugpushkey` to apply
the changes, effectively replaying the change that happened on the
server.

Replay of pushkey changes contains races. `hg pull` will often call
`listkeys` and pull down values. If the server has processed a `pushkey`
in the time it takes the mirror to process the `hg pull` (something that
is pretty much guaranteed to occur if a mirror gets out of sync), then a
mirror may receive pushkey data before the pushkey message is processed.
We will need to implement many more tests around this behavior. Testing
each widely-used pushkey namespace separately is probably a good idea.