vcsreplicator: tests for bookmark replication in a lagging client
`hg pull` will fetch bookmark data. If a mirror were to lag behind the
server, the mirror could be pulling changegroup data from a much newer
repository and it could potentially see bookmarks attached to changesets
it doesn't know of yet. This patch implements tests for this scenario.

The behavior of our replication is not ideal because the client can
sometimes get ahead of the log. However, there appears to be no harm in
this behavior. As the tests demonstrate, Mercurial recognizes when
remote bookmarks reference unknown changesets and simply ignores
applying them. And, when subsequent pushkey messages arrive and are
applied, they appear to no-op when they should no-op.

Long term, the non-ideal aspects of this problem are addressed by
bundle2 and/or replicating the bundles themselves and removing `hg pull`
from the equation. If we put all change data in the replication log
itself, we are in full control of the data Mercurial applies and thus
can better guarantee its behavior. But this doesn't appear to be an
issue (at least for bookmarks), so it isn't important to address now.