vcsreplicator: tests for phases replication in a lagging client
This is similar to the previously-added tests for bookmarks except it
tests phases as well. The tests are nearly identical and the behavior is
nearly the same. Again, Mercurial will pull pushkey data and temporarily
get ahead of the log. Again, this is acceptable and seems to result in
no bad behavior.