pushlog: try to write to a unified, Kafka-backed push log
Currently, vcsreplicator writes a unified log of push events that is
suitable for performing replication. There is a whole other class of
consumers that cares more about push events themselves.

This commit creates a new Kafka topic for holding data about just push
events (not replication info). It is logically a unified pushlog. The
pushlog has been modified to attempt to write into this unified pushlog
if it is available. If it isn't available, nothing new happens. If an
error happens, we record the error but don't abort.