mozreview: prevent concurrent autoland requests (bug 1220231); r=glob
Concurrent autoland requests on the same review request series to the same repository
are prevented using a global lock mechanism, which is based on memcached's add atomicity.
The expiration time of the lock is set to 24h to prevent situations where autoland errors
without reporting back and we get stuck with a invalid autoland lock id in the cache.