ansible/hg-ssh: check for writing events into Kakfa (bug 1126619); r=fubar
The replication log for hg has a no-op "heartbeat" message used for
testing whether the replication log is writable. This is used internally
so hg operations can fail fast when the replication log isn't available.
The message can also be used for monitoring. The sending of this message
is conveniently exposed via the `hg sendheartbeat` command.

This commit introduces a Nagios check that simply invokes `hg
sendheartbeat` and reports the results.

By running this check periodically in production, we'll verify the
replication log is online and writable so failures in the replication
log can hopefully be identified and fixed before a push comes in and
fails due to the log being unavailable.