vcsreplicator: fix test for full cluster restart
The old test said it would fail because a full cluster stop then start
would result in the cluster not coming back online. This assertion was
false: a cluster can come back online automatically as long as shutdown
was clean. The option we've enabled is the prevention of non ISR from
becoming leaders. We do not yet have a test for this scenario.

Change the test to verify a cluster is writable after a full restart.