ansible/hg-ssh: list all kafka servers in vcsreplicator config file
By listing all kafka servers, we ensure that if one goes down we can
continue talking to the cluster.

Before, we were only listing "localhost:9092." In this config, if the
Kafka server on the current host went down, pushes wouldn't work. This
is obviously not desirable.

A test demontrating that pushes work when hgssh's Kafka server is
offline has been added.