vcsreplicator: declare mapping of repo path to partition to write to
We now produce replication events into N partitions. The mapping from
repo to partition is deterministic so all events for a particular repo
are guaranteed to be in the same partition and thus strongly ordered
when consumed on the client.

We allocate 4 partitions for the main pool of repos on hg.mozilla.org. 1
partition for Try (because its replication events take much longer than
other repos) and 3 partitions for all user repos.

There are some hacky test changes here, mostly due to additional lines
of output and from not producing a heartbeat message into the same
partition as the init message. These should get addressed in subsequent
commits. 2 steps forward 1 step back.