vcsreplicator: store created time in message
We will be adding a monitor for how far consumers are out of sync with
the replication log. Kafka exposes a numeric message offset. But we also
want to know how far in time the replication log is behind. Since Kafka
doesn't store a timestamp in the message log, we need to store this
ourselves.