vcsreplicator: support for skipping over a message in the log
I've observed consumer processes getting stuck in an endless crash loop
in production. So far, I've only been able to reproduce this with
pulling into a corrupted repository.

There isn't really a terrific way to deal with this situation. But
having the consumer stall is very bad. So this commit provides a crude
tool to skip over the next message in the log. The theory being that if
a message causes a crash, we should stop trying to apply it. I'm
hesitant to do this automatically within the consumer daemon because it
could lead to badness. For now, let's have the monitors catch daemon
failures and require manual intervention if a message causes badness. We
can streamline the process later once we know what happens in the wild.