autoland: remove code paths for importing pull requests
This removes the code that supports importing pull requests from Autoland.
There is a lot of bitrot here, in particular this code still relies upon
the use of bugzilla cookies instead of api keys to work. Because of this
we can't test the pull request code paths when making changes to Autoland
which is dangerous. It's safer to remove it for now and add it back if we
plan to deploy support for pull requests.