ansible: install custom Pygments distribution (bug 1157502)
The Pygments project hasn't released an update in close to a year. It
contains a few patches that are interesting to us. But the most
important patch is to properly detect .jsm files as JavaScript so they
get syntax highlighted.

I have pushed the Pygments repository to
https://hg.mozilla.org/users/gszorc_mozilla.com/pygments. I cherry
picked the patch I submitted and was later merged into upstream onto the
commit that was tag 2.0.2. The Pygments package installed by this change
is thus Pygments 2.0.2 + my JSM fix
(https://bitbucket.org/birkenfeld/pygments-main/pull-requests/465/detect-jsm-files-as-javascript/diff).

Our deployments already pull version-control-tools from
https://hg.mozilla.org/, so no new operation dependency has been
introduced.