reviewboard: don't be so strict about verifying configuration settings (bug 1227715); r=mcote
Previously, the reviewboard server extension assumed that if it is
loaded that it needs to be configured. Attempting to run any `hg`
command on a machine with the extension in the system hgrc resulted in
an abort if the repository wasn't properly configured. This included
running `hg init`.

This commit loosens the requirements a bit. We introduce a new
configuration option defining the base directory of review
repos. The config validation only occurs for repositories under that
path. The extension will no longer interfere with `hg` invocations
unless they are being performed on declared review repositories.

The validation order and test output has changed slightly to make
the implementation easier.