autoland: pass commit descriptions to transplant and improve tests
This fixes a bad bug where the commit descriptions were not actually used when
tranplant is called.

This also improves the autoland tests so that similar bugs will be easier to
catch in the future. This removes the old unit tests which are not useful
because they depend upon the local test environment and will fail elsewhere.