mozreview: prevent RB from generating etags for review request pages (Bug 1221989). r=mdoglio
Review Board does not allow custom fields or extensions to effect etag
generation - Any server side modifications to what an extension would
display will be stale in a browser using caching.

To workaround this we remove etags Review Board would send to the client
as well as the HTTP_IF_NONE_MATCH header sent by the client (for a
select set of views we care about).

This will result in a bit of a performance hit but it shouldn't be too
noticeable as the really heavyweight requests like diff fragments (which
are fetched asynchronously from the diff viewer) will still be cached.