autoland: only push reviewed commits to landing repositories (bug 1228007) r=gps
If trysyntax is not present, we assume we are landing to an inbound
repository and only land reviewed commits. This changes the rebase command
to specify as a source the oldest reviewed commit.

This also changes hglib to run in non-interative mode and modifies the hg
identify command to specify '-r tip'.

Finally, the REST API is modified to enforce specifying either trysyntax
or commit_descriptions.