mozreview: upgrade to RB 2.5.2 (Bug 1231159). r=gps
Due to changes in Review Board's static media packaging code we now run
with DEBUG=False. This is required for Review Board to gather the
static media from our extensions. We also switch to using supported
means for installing and loading static media which may be slower but
works and should continue to work with future versions.