vcsreplicator: more robustly fetch message when checking lag
Before, we used the default timeout of 0.1s to fetch the message. This
is way too short. If this threshold were ever reached, we'd fail to
return a message and would attempt to destructure None into a tuple and
raise.

Speaking of that, we fix this error by checking the return value of
get_message(). If we fail to return a message, we report a high lag
time.