vcsreplicator: increase get message timeout
The default timeout is currently 0.1s. In production, we have 10
machines each polling 8 partitions. That's 80 consumer processes with
0.1s timeouts. Assuming no network latency, that can result in up to 800
requests per second! In reality, it appears to be much smaller - under
400 requests per second.

This commit increases the default timeout value to 2.0s. This reduces
the theoretical maximum requests per second to 40, which should be much
more manageable for the cluster.

Hopefully this makes the intermittent alerts go away. (My theory is the
servers weren't able to process incoming requests in an orderly manner.)