vcsreplicator: stop warning when unable to determine lag time
I feel dirty for having to do this. I'd rather figure out what is going
on. As it stands, this means we may not see legit alerts for a bit
longer (we need the message count to reach 5 to warn and that will take
2.5 minutes on average since we have 2 machines generating messages
every 60s as part of the message produce check).