firefoxtree: store firefox tree labels in a dedicated file; r=smacleod
Mercurial 3.3 has a "namespaces" API that allows namespaced labels to be
attached to Mercurial changesets. This API is a better version of the
"local tags" API, which firefoxtree is currently using. This patch
starts the series of converting firefoxtree to use the namespaces API.

We start the process by creating a standalone file that holds firefox
tree node mappings. The namespaces API doesn't do I/O for us, so we
have to roll some minimal code.

The .hg/firefoxtrees file is now written and read from when the
repository is loaded. It is currently a shadow copy of tags being
written to local tags. We will eventually transition off local tags
towards this new file.