firefoxtree: implement `hg fxheads` using new APIs; r=smacleod
Previously, we were printing tags. This was a holdover from the days
before {fxheads} was available. It has the unfortunate side-effect of
including non-tree tags in the output.

Change the template to use {fxheads} so only Firefox tree labels are
shown.

As part of this, limit to one output per unique node. Otherwise we will
print the same changeset twice if it has multiple fxtree labels defined.