firefoxtree: prune entries from localtags file; r=smacleod
We no longer write to this file. We need to prune existing entries from
it or they may interfere with existing use.

We could attempt to port values to the new namespace. But I feel it is
more work than it is worth: the proper values can be obtained by pulling
again, which is easy.