mozreview: show reviewer status in commit list (bug 1214614); r=smacleod
This adds a green background to the reviewer name when her last review has a ship-it.
Open issues are not taken into account yet, so a `fix it, then ship it` will result in a green
background on the reviewer name and a yellow icon for the whole commit status.