serverlog: make compatible with Mercurial 3.6; r=smacleod
I refactored how repository instances are used by hgweb in Mercurial
3.6. In the new world, a repo instance is attached to the request passed
into _runwsgi rather than the hgweb instance itself. This required
refactoring which function we wrap so we can access the repo instance
properly.

This commit does drop compatibility with Mercurial < 3.6. This should be
fine: Mozilla is the only consumer of this extension.