generate-hg-s3-bundles: only generate bzip2 bundles for release repos; r=smacleod
bzip2 bundles aren't used that much in the wild. They are most relevant
for release repos. So only generate them for those repos.

As part of this, we change the repos manifest file format to introduce
the concept of flags. Each value after the repo name gets turned into a
named argument to the generate_bundles function.