firefoxtree: check for firefoxtree attribute before accessing (bug 1234396); r=smacleod
I'm not sure what's causing the attribute to get lost. I can reproduce
it on my machine but I can't reproduce it in the test environment. Let's
work around it for now until a better solution can be implemented.

(The quick work around is important because this prevents people from
cloning.)