check_kafka_hg_heartbeat: rewrite to not use `hg sendheartbeat`
This check has been failing periodically in production for unknown
reasons. The solution for unwedging a partition is often to restart the
leader node. We desire to print broker information in check output to
help with diagnosing and correcting the issue. While we could put this
in `hg sendheartbeat`, it feels out of place there. `hg sendheartbeat`
was really just a convenience to avoid having to reimplement some code.

This commit reimplements `hg sendheartbeat` in the
check_kafka_hg_heartbeat Nagios check. By doing so, we can formulate
output to contain more details. We now print the failing partition and
its leader host in the failure output, allowing easier diagnosis of
failing partitions. There is still room to improve the output. This
will be done later.