mozreview: only include ship-it reviews in rewritten commit messages (Bug 1236725). r=dminor
Currently our commit rewriting will include users who did not give a
ship-it and users who gave a ship-it in the past but overrode it with a
non ship-it review. This does not match up with how the approval hook
functions and also adds reviewers who did not sign off on the code.

We should only include users who their most recent review has a ship-it.
While we're at it we should filter out the author as well. In the event
that the author is the only user to have given a ship-it, we will now
include an r=me to show they approved their own commit.