hgserver: replicate cloning via vcsreplicator (bug 1126153); r=smacleod
vcsreplicator doesn't currently intercept `hg clone` commands and
automatically generate replication events (it arguably should). So,
change how repository cloning is performed from `hg clone` to
`hg init` + `hg pull` + `hg replicatesync` so replication events are
written accordingly.

Also add in explicit testing that hgweb machines replicated the clone,
which was lacking before. This uncovered a bug in the lack of pash
creating per-user WSGI files. (This is done as an hourly CRON on the
servers to fill any gaps - but it should be done at repo creation time.)