reviewboard: declare jsonproto capability (bug 1234413); r=smacleod
Currently, we submit review requests to MozReview via a custom
Mercurial wire protocol command.

Back when we first wrote this code, the Mercurial environment on
Windows (MozillaBuild) was crippled because it was using a py2exe
distribution that didn't have various Python standard library modules
available. Calling web services easily and securely was not possible.
So, we decided to implement custom client-server communication via
the Mercurial wire protocol. The connection was already there and
usable, so why not. We invented our own line-based protocol for
data exchange.

The Windows environment no longer has a crippled Python environment.
Our original justification for leveraging the wire protocol has gone
away. The main remaining reason to use it is performance: the TCP
connection is already there, so latency is low.

In the near future, we will submit review requests from Git as well.
And, it would be a lot of effort to reimplement our custom wire
protocol exchange for the Git client.

So, the time is ripe to rewrite the client-server interaction as a
more traditional web service. This patch starts the process by laying
the groundwork to convert the existing data exchange from a line-based
protocol to JSON.

We introduce a new server and client capability that says each can
speak JSON. Neither side does yet. But you have to start somewhere.