reviewboard: convert reviewrepos command to JSON (bug 1234413); r=smacleod
We now begin the process of converting wire protocol commands to use
JSON. We start with the "reviewrepos" command, which is the simplest
wire protocol command we have.

As part of this, we establish a helper function to call a command and
obtain the JSON parsed response.

Now that we serve JSON, we should probably convert the repos list on
disk to JSON. We add a TODO to track doing this at some later time.