mozreview: cache reviewer to User lookups (bug 1229468); r=smacleod
The previous commit regressed the total number of BMO HTTP requests
because it stopped caching user lookups. This commit reintroduces that
caching.

The effect of this commit is the total number of HTTP requests to BMO is
reduced to the level they were at before the previous commit, or a net
reduction of 18 requests from the previous commit. Wall time didn't
decrease significantly. But I reckon this is because of the no latency,
no load environment. I have a feeling this series will have a dramatic
impact on production performance.