reviewboard: use built-in mechanism to not save cookies; r=smacleod
The hack around using a temporary file as a cookie manager was necessary
because older versions of RBTools didn't support not saving cookies. New
versions do. So remove the hack and use the built-in mechanism.

This function no longer needs to be a context manager. This will be
removed in a subsequent commit.