mozreview: create child review requests from batch API (bug 1229468); r=smacleod, dminor
The final step for moving the creation of review series into a unified
web API is moving the management of child review requests to the new
API, which this commit does.

This commit is rather large. Unfortunately, there is no easy way to
incrementally move this code since managing the state of child review
requests is so complicated and intertwined.

As part of migrating the code, the overall structure changed
substantially. This is because we're able to do things differently from
the context of the web api that we weren't from outside. For example, we
can query the user database directly instead of going through the search
interface.

The best way to review this change is probably to open a copy of
pushhooks.py in a separate window and read it alongside the new code
in batch_review_request.py.

Fortunately, our test coverage of review state is pretty robust. And, as
you can see, there are no test changes as a result of this refactor!
Presumably the new behavior is identical to the old behavior!

There is still room to optimize this code. For example, the old code
maintained a cache of reviewer lookups. This was dropped as part of the
refactor. It can and should be added in later.

The practical effect of this change is that review submission and
publishing now requires far fewer HTTP requests to the Review Board Web
API.

On a local test cluster, submitting and publishing a review series of 20
commits with a single reviewer required the following:

Before: 12.0s  134 BMO requests   158 RB requests
After:  10.0s  154 BMO requests    54 RB requests
Delta:  -2.0s  +20 HTTP requests -104 HTTP requests

(The HTTP request count includes requests during cluster startup, so it
is artificially high to start.)

The regression in BMO HTTP requests is almost certainly due to redundant
user lookups. This will be corrected later.