reviewboard: expose web command for submitting changesets for review; r=smacleod
Currently, we submit review requests to MozReview via a custom
Mercurial wire protocol command that (as of very recently) speaks
JSON.

In the near future, we will submit review requests from git-cinnabar
as well. While git-cinnabar does know how to speak the Mercurial wire
protocol and while it would be possible to have a custom Git command
call a Mercurial wire protocol command, it's just easier to use
JSON over HTTP.

This commit establishes a Mercurial web command that is called to
convert changesets into review requests. It calls the existing
code that powers the wire protocol command, so functionality should
be almost identical.

To prove the new API works, we convert the HTTP push flow to call the
web service instead of the wire protocol command. We don't have as much
test coverage of HTTP based pushes. But since the web API is basically a
proxy to existing code, everything should work the same.