mozreview: carry forward approval for autolanding (Bug 1239798) r=smacleod
Currently we require re-review on commits made by non-L3 users which have
been amended. This limits the utility of autoland as a replacement for the
manual Bugzilla "check-in needed" process. Instead, we will apply the same
criteria we use for L3 users for non-L3 users and create a warning in the
UI during the autoland approval process that commits have been amended and
may need further examination prior to landing.