ansible/mozreview-dev: make HG SSH server resemble production
The mozreview-dev Ansible configs aren't very robust and have little
resemblence to production. This commit attempts to address that.

We install Python 2.7 and configure virtualenvs for running pash. We
also install the latest version of pash on the server.

With this change, the reviewboard server extension actually works on the
dev server again.