reviewboard: report HTTP 500 errors without a stack (bug 1169664); r=smacleod
We had an uncaught APIError when submitting review requests. This caused
the server to print a nasty stack trace which has confused numerous
people. We catch the error and print the details.

We still have to fix the underlying issue causing the HTTP 500, which is
in the batch review request resource.