reviewboard: print more context from server error messages (bug 1169664); r=smacleod
We now display the error from the server.

The output for submitted review requests isn't perfect. This will be
addressed in a subsequent commit.

I removed the note about trying again because users shouldn't see this
error that often and I think most people will retry if the error doesn't
say to take another action.