autoland: prevent partial and incorrect landings (bug 1243276) r=gps
In Bug 1243276 we ended up with a partial landing of the wrong commits. The
landing tree was closed, but due to a bug with the treestatus code we did
not find out about this until push time, which was causing us to skip the
clean up code for non-public commits. I think this missing clean up along
with the fact that we pulled and rebased the same commits multiple times
got us in to a state where 'tip' no longer referred to the commits we
wanted to push.

This patch adds a strip of non-public commits to the beginning of any
transplant attempt. In addition, we add a check to make sure that the set
of incoming commit descriptions matches the set of outgoing commit
descriptions. This should prevent us from landing incorrect commits unless
we have somehow associated commit descriptions with the wrong commits.

I tested the new code manually by changing the handling of the encoding of
commit descriptions so they did not match after rewriting, but I not sure
we can test this automatically without having races between the test code
and autoland.