mozreview: Add model to track DiffSets from the hg server (Bug 1244448). r=mdoglio r=dminor r=gps
An important concept introduced with this model is that of a verified
DiffSet: Any DiffSet uploaded to MozReview by the hg server will have
a verification entry created for it (This operation will require the
special permission introduced). This will allow us to prevent
unauthorized diffs from being published, guaranteeing any diff shown
on MozReview reflects the commit backing it in hg.

Use of VerifiedDiffSet will come in following changes.