mozreview: require special permission to access `BatchReviewRequestResource.create` (Bug 1244448). r=mcote r=mdoglio r=gps
We now require the mozreview.verify_diffset permission to access the
batch review request creation resource. This means we can trust the
actions of this resource as coming from the hg server. This will be
important in future commits where we mark diffsets as verified using
the resource.

Also fixes the missing i in the spelling of privileged_rb_username and
privileged_rb_password.