mozreview: verify DiffSets during batch creation (Bug 1244448). r=mdoglio r=gps
DiffSetVerifications are created for each DiffSet that is uploaded
as part of batch review request creation. Actually checking these
verifications will come in a future change.

Additionally, providing a user with the verify_diffset permission is now
required in order to save a DiffSetVerification. While it wasn't
strictly necessary to add this, it means you're forced to think about
permissions whenever adding a place DiffSets get verified.