git-mozreview: git command for interacting with MozReview (bug 1153053); r=glandium, dminor
This commit establishes the `git mozreview` command which supports
submitting commits to MozReview for review.

The command consists of the "configure" and "push" sub-commands.
"configure" launches a wizard to configure the local Git repo to
use MozReview. It is a rough equivalent of `mach mercurial-setup`.

As part of configuration, we establish a "commit-msg" hook that
adds a unique identifier to commit messages. The intent of this
identifier is to track commits on MozReview. This will help
MozReview intelligently react to history rewrites. Unfortunately, we
don't yet use this identifier to track commits. But we have plans to
improve MozReview's tracking support in the weeks or months ahead,
so putting the tracking infrastructure in place now should make the
transition easier.

`git mozreview push` is the important command: it submits commits
to MozReview for review. The command first pushes commits to MozReview
using git-cinnabar then it makes a call to the appropriate web service
API to turn pushed commits (actually Mercurial changesets) into
review requests on MozReview. The workflow is modeled after the
existing Mercurial workflow.

MozReview-Commit-ID: 47EedQkLjeV